+++ test/coverlet.core.tests/Helpers/SourceRootTranslatorTests.cs
@@ -2,20 +2,20 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System.IO;
+using System.Runtime.InteropServices;
 using Coverlet.Core.Abstractions;
-using Coverlet.Tests.Xunit.Extensions;
+using Coverlet.Core.Helpers;
 using Moq;
 using Xunit;
 
-namespace Coverlet.Core.Helpers.Tests
+namespace Coverlet.Core.Tests.Helpers
 {
   public class SourceRootTranslatorTests
   {
-    [ConditionalFact]
-    [SkipOnOS(OS.Linux, "Windows path format only")]
-    [SkipOnOS(OS.MacOS, "Windows path format only")]
+    [Fact]
     public void Translate_Success()
     {
+      Assert.SkipUnless(RuntimeInformation.IsOSPlatform(OSPlatform.Windows), "Test requires Windows");
       string fileToTranslate = "/_/src/coverlet.core/obj/Debug/netstandard2.0/coverlet.core.pdb";
       var logger = new Mock<ILogger>();
       var assemblyAdapter = new Mock<IAssemblyAdapter>();
@@ -32,11 +32,10 @@ namespace Coverlet.Core.Helpers.Tests
       Assert.Equal(@"C:\git\coverlet\src\coverlet.core\obj\Debug\netstandard2.0\coverlet.core.pdb", translator.ResolveFilePath(fileToTranslate));
     }
 
-    [ConditionalFact]
-    [SkipOnOS(OS.Linux, "Windows path format only")]
-    [SkipOnOS(OS.MacOS, "Windows path format only")]
+    [Fact]
     public void TranslatePathRoot_Success()
     {
+      Assert.SkipUnless(RuntimeInformation.IsOSPlatform(OSPlatform.Windows), "Test requires Windows");
       var logger = new Mock<ILogger>();
       var assemblyAdapter = new Mock<IAssemblyAdapter>();
       assemblyAdapter.Setup(x => x.GetAssemblyName(It.IsAny<string>())).Returns("testLib");
@@ -51,11 +50,10 @@ namespace Coverlet.Core.Helpers.Tests
       Assert.Equal(@"C:\git\coverlet\", translator.ResolvePathRoot("/_/")[0].OriginalPath);
     }
 
-    [ConditionalFact]
-    [SkipOnOS(OS.Linux, "Windows path format only")]
-    [SkipOnOS(OS.MacOS, "Windows path format only")]
+    [Fact]
     public void TranslateWithDirectFile_Success()
     {
+      Assert.SkipUnless(RuntimeInformation.IsOSPlatform(OSPlatform.Windows), "Test requires Windows");
       var logger = new Mock<ILogger>();
       var assemblyAdapter = new Mock<IAssemblyAdapter>();
       assemblyAdapter.Setup(x => x.GetAssemblyName(It.IsAny<string>())).Returns("testLib");